requirements.txt
flask
flask-cors
onnxruntime
opencv-python-headless
numpy
Pillow
